requirements.txt
PySide6
python-vlc
pypresence
requests
